requirements.txt
# openai==1.12.0
python-dotenv==1.0.0
tqdm==4.66.1
pandas==2.1.4
google-generativeai==0.3.2
